LayoutTests/canvas/philip/tests/2d.path.stroke.scale2-expected.txt

< [index] >
2d.path.stroke.scale2
Stroke line widths are scaled by the current transformation matrix
References: 2d.path.transformation
Actual output:
Expected output:

Passed
